refactor: move define of types from common to config
